Make authorizer.isAllowed private
